+++ tailwind.config.cjs
@@ -1,4 +1,4 @@
-const kindlePreset = require('@oidanice/kindle-ui/preset')
+const kindlePreset = require('@oidanice/ink-ui/preset')
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
@@ -6,7 +6,7 @@ module.exports = {
   content: [
     './index.html',
     './src/**/*.{js,ts,jsx,tsx}',
-    './node_modules/@oidanice/kindle-ui/dist/**/*.{js,mjs}',
+    './node_modules/@oidanice/ink-ui/dist/**/*.{js,mjs}',
   ],
   theme: { extend: {} },
   plugins: [],